+++ Decimus.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,5 +1,5 @@
 {
-  "originHash" : "0e18a282738bac556786f777d0cc9d699b0c17faaa02a5c43533f19fbb9d4c90",
+  "originHash" : "e5d00282b90ec59408c99c73849de7d96c4d38bfbc0d6416d86320d2ff39d1db",
   "pins" : [
     {
       "identity" : "csv.swift",
