+++ Decimus/CaptureManager.swift
@@ -3,15 +3,27 @@ import UIKit
 import os
 
 public extension AVCaptureDevice {
+    /// An unsigned 64 bit probably unique identifier for a capture device.
     var id: UInt64 {
         .init(truncatingIfNeeded: uniqueID.hashValue)
     }
 }
 
+/// Represents the ability to receive camera frames.
 protocol FrameListener {
+    /// The queue that this object will receive camera frames on.
     var queue: DispatchQueue { get }
+
+    /// The device this listener is interested in receiving frames from.
     var device: AVCaptureDevice { get }
+
+    /// Optionally, the configuration that the listener wishes to receive frames in from ``device``.
+    /// This is a best-effort minimum value, callers MUST NOT assume the format of the received frames will match this.
     var codec: VideoCodecConfig? { get }
+
+    /// This function will be called with video frames from the target ``device``, executed on ``queue``.
+    /// - Parameter sampleBuffer: The camera frame.
+    /// - Parameter timestamp: The timestamp this frame was captured at.
     func onFrame(_ sampleBuffer: CMSampleBuffer, timestamp: Date)
 }
 
@@ -23,16 +35,23 @@ fileprivate extension FrameListener {
     }
 }
 
+/// Possible errors raised by ``CaptureManager``.
 enum CaptureManagerError: Error {
+    /// `AVCaptureMultiCamSession` is not supported on this platform.
     case multicamNotSuported
+    /// The requested operation is invalid in this state. (E.g start when started).
     case badSessionState
+    /// The requested operation targeted a device is not registered.
     case missingInput(AVCaptureDevice)
+    /// Failed to add the target device.
     case couldNotAdd(AVCaptureDevice)
+    /// ``CaptureManager`` should not be used to manage audio devices. See ``DecimusAudioEngine``.
     case noAudio
+    /// The requested operation MUST be called from the main thread, and was not.
     case mainThread
 }
 
-/// Manages local media capture.
+/// Manages local video capture.
 class CaptureManager: NSObject, AVCaptureVideoDataOutputSampleBufferDelegate {
     private static let logger = DecimusLogger(CaptureManager.self)
 
@@ -57,6 +76,9 @@ class CaptureManager: NSObject, AVCaptureVideoDataOutputSampleBufferDelegate {
     private var pressureObservations: [AVCaptureDevice: NSObjectProtocol] = [:]
     private let bootDate: Date
 
+    /// Create a new ``CaptureManager``.
+    /// - Parameter metricsSubmitter: Optionally, a submitter to collect/submit metrics through.
+    /// - Parameter granularMetrics: Collect granular metrics when a submitter is present, at a potential performance penalty.
     init(metricsSubmitter: MetricsSubmitter?, granularMetrics: Bool) throws {
         guard AVCaptureMultiCamSession.isMultiCamSupported else {
             throw CaptureManagerError.multicamNotSuported
@@ -74,21 +96,30 @@ class CaptureManager: NSObject, AVCaptureVideoDataOutputSampleBufferDelegate {
         super.init()
     }
 
+    /// Get a list of all managed devices.
+    /// - Throws: ``CaptureManagerError/mainThread``. Must be called on the main thread.
     func devices() throws -> [AVCaptureDevice] {
         guard Thread.isMainThread else { throw CaptureManagerError.mainThread }
         return Array(connections.keys)
     }
 
+    /// Get the subset of ``devices()`` that are active.
+    /// - Throws: ``CaptureManagerError/mainThread``. Must be called on the main thread.
     func activeDevices() throws -> [AVCaptureDevice] {
         guard Thread.isMainThread else { throw CaptureManagerError.mainThread }
         return Array(try connections.keys.filter { try !isMuted(device: $0) })
     }
 
+    /// Is the given device is already registered to the manager?
+    /// - Returns: True if managed.
+    /// - Throws: ``CaptureManagerError/mainThread``. Must be called on the main thread.
     func usingInput(device: AVCaptureDevice) throws -> Bool {
         guard Thread.isMainThread else { throw CaptureManagerError.mainThread }
         return inputs[device] != nil
     }
 
+    /// Start capturing video from all target devices.
+    /// - Throws: ``CaptureManagerError/mainThread``. Must be called on the main thread. ``CaptureManagerError/badSessionState`` if already running.
     func startCapturing() throws {
         guard Thread.isMainThread else { throw CaptureManagerError.mainThread }
         guard !session.isRunning else {
@@ -108,8 +139,7 @@ class CaptureManager: NSObject, AVCaptureVideoDataOutputSampleBufferDelegate {
         }
     }
 
-    @Sendable
-    private nonisolated func onStartFailure(notification: Notification) {
+    private func onStartFailure(notification: Notification) {
         guard let error = notification.userInfo?[AVCaptureSessionErrorKey] as? AVError else {
             Self.logger.error("AVCaptureSession failed for unknown reason")
             return
@@ -117,6 +147,8 @@ class CaptureManager: NSObject, AVCaptureVideoDataOutputSampleBufferDelegate {
         Self.logger.error("AVCaptureSession failure: \(error.localizedDescription)")
     }
 
+    /// Stop capturing media.
+    /// - Throws: ``CaptureManagerError/mainThread``. Must be called on the main thread. ``CaptureManagerError/badSessionState`` if already stopped.
     func stopCapturing() throws {
         guard Thread.isMainThread else { throw CaptureManagerError.mainThread }
         guard session.isRunning else {
@@ -125,6 +157,10 @@ class CaptureManager: NSObject, AVCaptureVideoDataOutputSampleBufferDelegate {
         self.session.stopRunning()
     }
 
+    /// Mute/unmute the target device.
+    /// - Parameter device: The device to toggle mute status on.
+    /// - Parameter toggled: Callback with the new current state.
+    /// - Throws: ``CaptureManagerError/mainThread``. Must be called on the main thread.
     func toggleInput(device: AVCaptureDevice, toggled: @escaping (Bool) -> Void) throws {
         guard Thread.isMainThread else { throw CaptureManagerError.mainThread }
         guard let connection = self.connections[device] else { fatalError() }
@@ -236,6 +272,9 @@ class CaptureManager: NSObject, AVCaptureVideoDataOutputSampleBufferDelegate {
         self.multiVideoDelegate[device] = [listener]
     }
 
+    /// Add a listener for frame callbacks, adding the target device if not already.
+    /// - Parameter listener: Receiver of video frames.
+    /// - Throws: ``CaptureManagerError/mainThread``. Must be called on the main thread.
     func addInput(_ listener: FrameListener) throws {
         guard Thread.isMainThread else { throw CaptureManagerError.mainThread }
         Self.logger.info("Adding capture device: \(listener.device.localizedName)")
@@ -249,6 +288,9 @@ class CaptureManager: NSObject, AVCaptureVideoDataOutputSampleBufferDelegate {
         try addCamera(listener: listener)
     }
 
+    /// Remove a listener for frame callbacks, removing the target device if no other listeners targetting that device are left.
+    /// - Parameter listener: Receiver of video frames.
+    /// - Throws: ``CaptureManagerError/mainThread`` if called on a thread other than Main. ``CaptureManagerError/missingInput(_:)`` if the device is not already tracked.
     func removeInput(listener: FrameListener) throws {
         guard Thread.isMainThread else { throw CaptureManagerError.mainThread }
 
@@ -286,6 +328,10 @@ class CaptureManager: NSObject, AVCaptureVideoDataOutputSampleBufferDelegate {
         Self.logger.info("Removing input for \(device.localizedName)")
     }
 
+    /// Query the mute state of the target device.
+    /// - Parameter device: The device to query.
+    /// - Returns: True if currently muted.
+    /// - Throws: ``CaptureManagerError/mainThread`` if called on a thread other than Main. ``CaptureManagerError/missingInput(_:)`` if the device is not already tracked.
     func isMuted(device: AVCaptureDevice) throws -> Bool {
         guard Thread.isMainThread else { throw CaptureManagerError.mainThread }
         guard let connection = connections[device] else {
@@ -294,6 +340,9 @@ class CaptureManager: NSObject, AVCaptureVideoDataOutputSampleBufferDelegate {
         return !connection.isEnabled
     }
 
+    /// Add a preview view for a device.
+    /// - Parameter device: The device to add a preview for.
+    /// - Parameter preview: The preview layer to hook to the device.
     func addPreview(device: AVCaptureDevice, preview: AVCaptureVideoPreviewLayer) throws {
         guard Thread.isMainThread else { throw CaptureManagerError.mainThread }
         guard let connection = connections[device] else {
@@ -314,6 +363,7 @@ class CaptureManager: NSObject, AVCaptureVideoDataOutputSampleBufferDelegate {
         return subscribers
     }
 
+    /// `AVCaptureVideoDataOutputSampleBufferDelegate` camera frame callback.
     func captureOutput(_ output: AVCaptureOutput,
                        didOutput sampleBuffer: CMSampleBuffer,
                        from connection: AVCaptureConnection) {
@@ -345,6 +395,7 @@ class CaptureManager: NSObject, AVCaptureVideoDataOutputSampleBufferDelegate {
         }
     }
 
+    /// `AVCaptureVideoDataOutputSampleBufferDelegate` dropped frame callback.
     func captureOutput(_ output: AVCaptureOutput,
                        didDrop sampleBuffer: CMSampleBuffer,
                        from connection: AVCaptureConnection) {
@@ -411,6 +462,7 @@ class CaptureManager: NSObject, AVCaptureVideoDataOutputSampleBufferDelegate {
 
 #if !os(tvOS)
 extension UIDeviceOrientation {
+    /// Get Decimus' representation of this orientation.
     var videoOrientation: DecimusVideoRotation {
         switch self {
         case .portrait:
