+++ Decimus/DecimusVideoFrame.swift
@@ -1,14 +1,23 @@
 import AVFoundation
 
+/// A video frame and all related appliaction metadata as used in Decimus.
 class DecimusVideoFrame {
+    /// Video data samples and attachments (typically one).
     let samples: [CMSampleBuffer]
+    /// The MoQ group ID for this frame.
     let groupId: UInt64
+    /// The MoQ object ID for this frame.
     let objectId: UInt64
+    /// The sequence number for this frame.
     let sequenceNumber: UInt64?
+    /// If present, the expected FPS (1/duration) of this frame's stream.
     let fps: UInt8?
+    /// If present, the orientation of this video frame.
     let orientation: DecimusVideoRotation?
+    /// If present and true, this frame is vertically mirrored.
     let verticalMirror: Bool?
 
+    /// Create a new video frame from its parts.
     init(samples: [CMSampleBuffer],
          groupId: UInt64,
          objectId: UInt64,
@@ -25,7 +34,9 @@ class DecimusVideoFrame {
         self.verticalMirror = verticalMirror
     }
 
-    /// Create a video frame from a deep copy of the provided frame and its data.
+    /// Create a video frame by deep copying the provided frame and its data.
+    /// - Parameter copy: The frame to deep copy.
+    /// - Throws: No data, or failure to create destination memory blocks.
     init(copy: DecimusVideoFrame) throws {
         // Deep copy all sample data blocks.
         var samples: [CMSampleBuffer] = []
