--- Decimus/Models/SubscriptionError.swift
@@ -1,4 +1,0 @@
-enum SubscriptionError: Int32 {
-    case none = 0
-    case noDecoder
-}
