+++ Decimus/VideoJitterBuffer.swift
@@ -8,16 +8,25 @@ import Atomics
 /// A very simplified jitter buffer designed to contain compressed video frames in order.
 class VideoJitterBuffer {
 
+    /// Jiitter buffer configuration.
     struct Config: Codable {
+        /// The mode of operation.
         var mode: Mode = .none
+        /// The initial target & minimum depth frames will be delayed for.
         var minDepth: TimeInterval = 0.2
+        /// The maximum storage capacity of the jitter buffer.
         var capacity: TimeInterval = 2
     }
 
+    /// Possible modes of jitter buffer usage.
     enum Mode: CaseIterable, Identifiable, Codable {
+        /// Use a PID controller to maintain target depth.
         case pid
+        /// Attempt to schedule individual frames to be played out according to their timestamp.
         case interval
+        /// Delegate timing control to `AVSampleBufferDisplayLayer`
         case layer
+        /// No jitter buffer, play frames syncronously on receive.
         case none
         var id: Self { self }
     }
@@ -31,11 +40,11 @@ class VideoJitterBuffer {
     private var lastSequenceSet = ManagedAtomic<Bool>(false)
 
     /// Create a new video jitter buffer.
-    /// - Parameter fullTrackName The namespace of the video this buffer is used for, for identification purposes.
-    /// - Parameter metricsSubmitter Optionally, an object to submit metrics through.
-    /// - Parameter sort True to actually sort on sequence number, false if they're already in order.
-    /// - Parameter minDepth Fixed initial delay in seconds.
-    /// - Parameter capacity Capacity in buffers / elements.
+    /// - Parameter fullTrackName: The namespace of the video this buffer is used for, for identification purposes.
+    /// - Parameter metricsSubmitter: Optionally, an object to submit metrics through.
+    /// - Parameter sort: True to actually sort on sequence number, false if they're already in order.
+    /// - Parameter minDepth: Fixed initial & target delay in seconds.
+    /// - Parameter capacity: Capacity in number of buffers / elements.
     init(fullTrackName: FullTrackName,
          metricsSubmitter: MetricsSubmitter?,
          sort: Bool,
@@ -92,7 +101,8 @@ class VideoJitterBuffer {
 
     /// Write a video frame into the jitter buffer.
     /// Write should not be called concurrently with another write.
-    /// - Parameter videoFrame The sample to attempt to sort into the buffer.
+    /// - Parameter videoFrame: The sample to attempt to sort into the buffer.
+    /// - Throws: Buffer is full, or video frame is older than last read.
     func write(videoFrame: DecimusVideoFrame, from: Date) throws {
         // Check expiry.
         if let thisSeq = videoFrame.sequenceNumber,
@@ -113,7 +123,8 @@ class VideoJitterBuffer {
     }
 
     /// Attempt to read a frame from the front of the buffer.
-    /// - Returns Either the oldest available frame, or nil.
+    /// - Parameter from: The timestamp of this read operation. This is an optimization to reduce the number of now() calculations, and has no bearing on jitter buffer behaviour.
+    /// - Returns: Either the oldest available frame, or nil.
     func read(from: Date) -> DecimusVideoFrame? {
         let depth: TimeInterval? = self.measurement != nil ? self.buffer.duration.seconds : nil
 
@@ -142,23 +153,23 @@ class VideoJitterBuffer {
     }
 
     /// Get the CMBuffer at the front of the buffer without removing it.
-    /// - Returns The head of the buffer, if any.
+    /// - Returns: The head of the buffer, if any.
     func peek() -> CMBuffer? {
         self.buffer.head
     }
 
     /// Get the current depth of the queue (sum of all contained durations).
-    /// - Returns Duration in seconds.
+    /// - Returns: Duration in seconds.
     func getDepth() -> TimeInterval {
         self.buffer.duration.seconds
     }
 
     /// Calculate the estimated time interval until this frame should be rendered.
-    /// - Parameter frame The frame to calculate the wait time for.
-    /// - Parameter from The time to calculate the time interval from.
-    /// - Parameter offset Offset from the start point at which media starts.
-    /// - Parameter since The start point of the media timeline.
-    /// - Returns The time to wait, or nil if no estimation can be made. (There is no next frame).
+    /// - Parameter frame: The frame to calculate the wait time for.
+    /// - Parameter from: The time to calculate the time interval from.
+    /// - Parameter offset: Offset from the start point at which media starts.
+    /// - Parameter since: The start point of the media timeline.
+    /// - Returns: The time to wait, or nil if no estimation can be made. (There is no next frame).
     func calculateWaitTime(frame: DecimusVideoFrame,
                            from: Date,
                            offset: TimeInterval,
@@ -179,10 +190,10 @@ class VideoJitterBuffer {
     }
 
     /// Calculate the estimated time interval until the next frame should be rendered.
-    /// - Parameter from The time to calculate the time interval from.
-    /// - Parameter offset Offset from the start point at which media starts.
-    /// - Parameter since The start point of the media timeline.
-    /// - Returns The time to wait, or nil if no estimation can be made. (There is no next frame).
+    /// - Parameter from: The time to calculate the time interval from.
+    /// - Parameter offset: Offset from the start point at which media starts.
+    /// - Parameter since: The start point of the media timeline.
+    /// - Returns: The time to wait, or nil if no estimation can be made. (There is no next frame).
     func calculateWaitTime(from: Date,
                            offset: TimeInterval,
                            since: Date = .init(timeIntervalSince1970: 0)) -> TimeInterval? {
