--- Decimus/WrappedOptional.swift
@@ -1,6 +1,0 @@
-class WrappedOptional<T> {
-    var value: T?
-    init(_ value: T?) {
-        self.value = value
-    }
-}
